+++ dts/math-and-unit-parser.d.ts
@@ -1,0 +1,3 @@
+declare module 'math-and-unit-parser';
+
+declare function MathParser(input: string): MathParser;