+++ src/functions.ts
@@ -1,11 +1,10 @@
 // Default built-in functions
 
-import NovaSheets from './index';
+import NovaSheets from './novasheets';
 import { Constants, CustomFunction } from './common';
 
 function addBuiltInFunctions({ constants }: { constants: Constants }): CustomFunction[] {
-    const novasheets: NovaSheets = new (require('./index'))(); // ???
-    // const novasheets: NovaSheets = new NovaSheets();
+    const novasheets: NovaSheets = new NovaSheets();
     const escapeRegex = (str: string): string => str.replace(/[.*+?^/${}()|[\]\\]/g, '\\$&');
     const strim = (str: string): string => str.toString().replace(/^\s*(.+?)\s*$/, '$1').replace(/\s+/g, ' ');
     const r = String.raw;
@@ -338,4 +337,4 @@ function addBuiltInFunctions({ constants }: { constants: Constants }): CustomFun
 
 }
 
-export = addBuiltInFunctions;
+export default addBuiltInFunctions;