+++ src/index.ts
@@ -1,3 +1,6 @@
+// CommonJS interop //
+// Do not import this file in module form; either `require('index')` or `import _ from 'novasheets'`.
+
 import NovaSheets from './novasheets';
 
 export = NovaSheets;